[FIX] pos_loyalty: prevent error when archiving a program
Before this commit, archiving a loyalty program while it was actively
applied to an order could lead to errors.

Specifically, if a PoS order had an active loyalty program applied, and
that program was subsequently archived from the backend, reloading the
PoS interface would cause issues or prevent proper functionality due to
the program's sudden unavailability.

opw-4941044

closes odoo/odoo#218814

Signed-off-by: David Monnom (moda) <moda@odoo.com>